requirements.txt
django==2.1.7
djangorestframework==3.9.1
django-webpack-loader==0.6.0
whitenoise==4.1.2
